refactor(external): dto -> multivalueMap 변경 로직 client로 이관
- 외부 로직 관련 dto 패키지 경로를 `dto/external`로 이관
