+++ src/main/java/gift/dto/external/KakaoTokenRequest.java
@@ -1,0 +1,13 @@
+package gift.dto.external;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public record KakaoTokenRequest(
+        String grantType,
+        String clientId,
+        String redirectUri,
+        String code
+) {
+}
